requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
isodate==0.6.1
python-telegram-bot==20.7
lxml==4.9.3
